Create Mini-Project.C.git
Ramesh is working as an Anganwadi worker for an organization. He got a duty to collect the
vaccination details for COVID of all people and give them a vaccination ticket which would be
helpful for the people at the place of vaccination checking. Write a C program to help him to
collect the details and print a vaccination ticket for them.